requirements.txt
asgiref==3.6.0
attrs==22.2.0
cattrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docstring-to-markdown==0.12
filelock==3.10.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
jedi==0.18.2
jedi-language-server==0.40.0
Jinja2==3.1.2
lsprotocol==2023.0.0a1
MarkupSafe==2.1.2
mock-open==1.4.0
openapi-codec==1.3.2
parso==0.8.3
Pillow==9.4.0
platformdirs==3.1.1
psycopg2==2.9.5
pydantic==1.10.6
pygls==1.0.1
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
simplejson==3.18.4
sqlparse==0.4.3
typeguard==2.13.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
